--- a/Project.toml
+++ b/Project.toml
@@ -1,7 +1,7 @@
 name = "NumericalRelaxation"
 uuid = "82303a6c-a702-44d7-938a-15dad61d8a03"
 authors = ["Maximilian Köhler <maximilian.koehler@ruhr-uni-bochum.de>", "Timo Neumeier <timo.neumeier@math.uni-augsburg.de>", "Jan Sebastian Melchior <Jan.Melchior-a1m@ruhr-uni-bochum.de>", "Contributors"]
-version = "1.0.2"
+version = "1.0.3"
 
 [deps]
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"