see https://discourse.julialang.org/t/splatting-arguments-causes-30x-slow-down/16964

### main 
```
julia> @benchmark convexify(convexification,buffer,W,Tensor{2,1}((2.0,)),material,state)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  171.092 μs …   3.080 ms  ┊ GC (min … max): 0.00% … 92.88%
 Time  (median):     178.677 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   193.059 μs ± 170.839 μs  ┊ GC (mean ± σ):  6.46% ±  6.76%

    ▃▅▆▆▇██▇▆▅▅▃▃▃▂▂▁▁           ▁▁▁▁▁                          ▂
  ▆█████████████████████▇▇▅▆▇▆▇████████▇▇▆▇▅▇▅▅▄▄▃▂▄▅▅▅▅▂▄▄▄▅▅▆ █
  171 μs        Histogram: log(frequency) by time        226 μs <

 Memory estimate: 245.48 KiB, allocs estimate: 10295.

```
### PR
```
julia> @benchmark convexify(convexification,buffer,W,Tensor{2,1}((2.0,)),material,state)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  77.503 μs … 111.495 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     78.358 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   78.459 μs ±   1.481 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

            ▁▁▁▅██▃
  ▂▂▃▃▃▃▄▅▇████████▇▄▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▁▂▂▂▁▁▁▂▁▁▂▂▁▂▂▁▁▂▂▂ ▃
  77.5 μs         Histogram: frequency by time         81.2 μs <

 Memory estimate: 288 bytes, allocs estimate: 5.
```